+++ saleor/graphql/page/resolvers.py
@@ -1,10 +1,8 @@
-import graphene
-
 from ...core.tracing import traced_resolver
 from ...page import models
 from ..core.utils import from_global_id_or_error
 from ..core.validators import validate_one_of_args_is_in_query
-from .types import PageType
+from .types import Page
 
 
 @traced_resolver
@@ -15,7 +13,7 @@ def resolve_page(info, global_page_id=None, slug=None):
     if slug is not None:
         page = models.Page.objects.visible_to_user(user).filter(slug=slug).first()
     else:
-        _type, page_pk = from_global_id_or_error(global_page_id)
+        _type, page_pk = from_global_id_or_error(global_page_id, Page)
         page = models.Page.objects.visible_to_user(user).filter(pk=page_pk).first()
     return page
 
@@ -26,9 +24,8 @@ def resolve_pages(info, **_kwargs):
     return models.Page.objects.visible_to_user(user)
 
 
-@traced_resolver
-def resolve_page_type(info, global_page_type_id):
-    return graphene.Node.get_node_from_global_id(info, global_page_type_id, PageType)
+def resolve_page_type(id):
+    return models.PageType.objects.filter(id=id).first()
 
 
 @traced_resolver