+++ saleor/graphql/payment/schema.py
@@ -2,10 +2,11 @@ import graphene
 
 from ...core.permissions import OrderPermissions
 from ..core.fields import FilterInputConnectionField
+from ..core.utils import from_global_id_or_error
 from ..decorators import permission_required
 from .filters import PaymentFilterInput
 from .mutations import PaymentCapture, PaymentInitialize, PaymentRefund, PaymentVoid
-from .resolvers import resolve_payments
+from .resolvers import resolve_payment_by_id, resolve_payments
 from .types import Payment
 
 
@@ -25,11 +26,12 @@ class PaymentQueries(graphene.ObjectType):
 
     @permission_required(OrderPermissions.MANAGE_ORDERS)
     def resolve_payment(self, info, **data):
-        return graphene.Node.get_node_from_global_id(info, data.get("id"), Payment)
+        _, id = from_global_id_or_error(data["id"], Payment)
+        return resolve_payment_by_id(id)
 
     @permission_required(OrderPermissions.MANAGE_ORDERS)
-    def resolve_payments(self, info, query=None, **_kwargs):
-        return resolve_payments(info, query)
+    def resolve_payments(self, info, **_kwargs):
+        return resolve_payments(info)
 
 
 class PaymentMutations(graphene.ObjectType):